+++ src/pages/board/index.jsx
@@ -1,15 +1,14 @@
 import React from "react"
-import  './mainboard.scss';
 
 const MainBoard= () => {
   return(
         <div>
 
-                <div className="mainBoard">
+          <div className="mainBoard">
                         <header className="headerBoard">
                                 <input className="nameBoard"  type="text" placeholder=" Tittle"/>
                                 <a  href="/"> <img className="iconsBoard" src="..\img\star-regular.png" alt=""/></a>
-                                <a  href="/">Espacio de trabajo de Trello</a>   
+                                <a  href="/">Espacio de trabajo de Trello</a>
                                 <a  href="/"><img className="iconsBoard" src="..\img\users-solid.png" alt="" /> Workspace visible</a>
                                 <a  href="/"><img className="iconsBoard" src="..\img\share-square-solid.png" alt="" />  Share</a>
                                 <a  href="/"><img className="iconsBoard" src="..\img\filter-solid.png" alt="" />  Filter</a>
