+++ src/pages/home/index.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import  './home.scss';
+
