+++ src/pages/login/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 
-import  './login.scss';
 
 function LogIn() {
   return(
