+++ src/styles/components/backgroundBoard.scss
@@ -1,0 +1,101 @@
+$background-image1: url('../../../public/img/backgroundImage1.jpg');
+$background-image2: url('../../../public/img/backgroundImage2.jpg');
+$background-image3: url('../../../public/img/backgroundImage3.jpg');
+$background-image4: url('../../../public/img/backgroundImage4.jpg');
+
+$background-blue: rgb(0, 121, 191);
+$backgound-orange: rgb(210, 144, 52);
+$background-green: rgb(81, 152, 57);
+$background-red: rgb(176, 70, 50);
+
+.board__ul {
+    list-style: none;
+    margin: 0;
+    padding: 0; 
+    display: flex;
+    justify-content: space-between;
+
+    &:first-child {
+        padding-bottom: 8px;
+    }
+}
+
+.board__li {
+    display: list-item;
+    text-align: -webkit-match-parent;
+    width: 64px;
+    height: 40px;
+
+    & button {
+        align-items: center;
+        background-position: center center;
+        background-size: cover;
+        border-radius: 3px;
+        box-shadow: none;
+        color: rgba(255, 255, 255, 0.4);
+        display: flex;
+        height: 100%;
+        justify-content: center;
+        margin: 0;
+        min-height: 0;
+        outline: 0;
+        padding: 0;
+        position: relative;
+        width: 100%;
+        cursor: pointer;   
+        
+        &:hover {            
+            bottom: 0;
+            display: block;
+            left: 0;
+            right: 0;
+            top: 0;
+            border-radius: 3px;
+            opacity: 0.75;
+            transition: .4s all;
+        } 
+          
+    }
+
+    &__button {        
+        font-family: 'Helvetica Neue', sans-serif;
+        font-size: 14px;
+        font-weight: 400;
+        box-sizing: border-box;        
+        border: none;   
+        text-decoration: none;         
+    }    
+    
+}
+
+.button1 {
+    background-image: $background-image1;
+}
+.button2 {
+    background-image: $background-image2;
+}
+.button3 {
+    background-image: $background-image3;
+}
+
+.button4 {
+    background-image: $background-image4;
+}
+
+
+.button5 {
+    background-color: $background-blue;
+}
+
+.button6 {
+    background-color: $backgound-orange;
+}
+
+.button7 {
+    background-color: $background-green;
+}
+
+.button8 {
+    background-color: $background-red;
+}
+
