+++ src/styles/components/createBoard.scss
@@ -1,0 +1,262 @@
+
+.board {
+    color: #172b4d;
+    font-family: 'Helvetica Neue', sans-serif;
+    font-size: 14px;
+    line-height: 20px;
+    font-weight: 400;
+    background-color: #ffffff;
+    border-radius: 3px;
+    box-shadow: 0 8px 16px -4px rgb(9 30 66 / 25%), 0 0 0 1px rgb(9 30 66 / 8%);
+    box-sizing: border-box;    
+    left: 4px;
+    overflow: hidden;    
+    position: fixed;
+    top: 48px;
+    width: 304px;
+    will-change: top, left; 
+
+    &__header {
+        margin-bottom: 8px;
+        padding: 0 12px;
+        position: relative;
+        text-align: center;
+        display: grid;
+        align-items: center;
+        grid-template-columns: 12px 1fr 12px;
+
+        &__div {
+            border-bottom: 1px solid rgba(9, 30, 66, 0.13);
+            color: #5e6c84;
+            display: block;
+            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Noto Sans', 'Ubuntu', 'Droid Sans', 'Helvetica Neue', sans-serif;
+            font-size: 16px;    
+            font-weight: 400;
+            grid-column: 1 / span 3;
+            grid-row: 1; 
+            height: 40px;
+            line-height: 40px;
+            margin: 0;
+            overflow: hidden;
+            padding: 0 32px;
+            position: relative;
+            text-overflow: ellipsis;
+            white-space: nowrap; 
+        }
+
+        &__button {
+            background: transparent;
+            border: none;
+            color: #6b778c;
+            cursor: pointer;
+            font-size: 16px;
+            grid-column: 3;
+            grid-row: 1;        
+            padding: 0;       
+            z-index: 2; 
+        
+            &:hover {
+                color: black;
+            }
+        }
+    }
+
+    &__section {
+        overflow-x: hidden;
+        overflow-y: auto;
+        padding: 12px;
+        max-height: 554px;  
+        top: 0;              
+
+        &__div {            
+            align-items: center; 
+            background-image: url("https://images.unsplash.com/photo-1657089343320-cc64d183de1b?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=Mnw3MDY2fDB8MXxjb2xsZWN0aW9ufDF8MzE3MDk5fHx8fHwyfHwxNjU3MjEyNDMy&ixlib=rb-1.2.1&q=80&w=400");
+            background-position: center center;
+            background-size: cover;
+            border-radius: 3px;
+            box-shadow: 0 7px 15px rgb(0 0 0 / 15%);
+            display: flex;            
+            height: 120px;
+            justify-content: center;  
+            margin-left: 33px;         
+            padding-bottom: 8px;
+            width: 200px;
+            margin-bottom: 20px;
+        }
+
+        &__image {
+            height: 80px;
+            width: 120px;
+            justify-content: center;
+            margin-top: 8px;
+        }
+    
+
+        &__background {
+        padding-bottom: 4px;
+
+            &__label {
+                color: #5e6c84;
+                cursor: default;
+                display: block;
+                font-size: 12px;
+                font-weight: 700;
+                line-height: 16px;
+                margin-bottom: 8px;
+                margin-top: 12px;
+            }
+        }
+        
+        &__form {
+            display: block;
+            margin-top: 0em;
+        }
+
+        &__label {
+            color: #5e6c84;
+            font-size: 12px;
+            line-height: 16px;
+            font-weight: 700;
+            display: flex;
+            margin-bottom: 8px;
+            margin-top: 12px;
+            cursor: default;
+        }
+
+        &__span {
+            color: #eb5a46;
+            margin-left: 2px;
+            font-weight: bold;
+
+            &--2 {
+                color: #172b4d;
+                display: flex;
+                font-size: 14px;
+                font-weight: 400;
+                flex-direction: row;
+                text-align: center;
+                margin-right: 8px;
+                margin-bottom: 4px;
+                margin-top: 12px;
+                cursor: default;
+            }
+        }
+
+        &__input {            
+            background-color: #ebecf0;
+            outline: none;
+            border: none;
+            box-sizing: border-box;
+            color: #172b4d;
+            font-family: 'Helvetica Neue', sans-serif;
+            font-size: 14px;
+            line-height: 20px;
+            font-weight: 400;
+            transition-duration: 85ms;
+            transition-timing-function: ease;
+            border-radius: 3px;
+            padding: 8px 12px;
+            box-shadow: inset 0 0 0 2px #dfe1e6;
+            background-color: #fafbfc;
+            width: 100%;
+            margin-bottom: 0;
+
+            &:hover {
+                box-shadow: inset 0 0 0 2px #eb5a46;
+                background-color: #ebecf0;
+            }
+        }
+
+        &__button {
+            background-color: #0079bf;
+            box-shadow: none;
+            border: none;
+            color: #ffffff;
+            width: 100%;
+            margin-bottom: 10px;
+            margin-top: 16px;
+            font-size: 14px;
+            line-height: 20px;
+            font-weight: 400;
+            box-sizing: border-box;
+            display: inline-flex;
+            align-items: center;
+            justify-content: center;
+            border-radius: 3px;
+            cursor: pointer;
+            padding: 6px 12px;
+            text-decoration: none;
+            white-space: normal;
+
+            &:hover {
+                background-color: #026aa7;
+                box-shadow: none;
+                border: none;
+                color: #ffffff;
+            }
+
+            &:disabled {
+                background-color: rgba(9, 30, 66, 0.04);
+                box-shadow: none;
+                border: none;
+                color: #a5adba;
+                cursor: not-allowed;
+            }
+        }   
+        
+        &__foooter {
+            color: #5e6c84;
+            margin-bottom: 8px;
+        }
+
+        &__text {
+            text-align: left;
+            font-family: 'Helvetica Neue', sans-serif;
+            font-size: 12px;            
+        }
+
+        &__a {                     
+            color: #5e6c84;
+            cursor: pointer;
+            font-size: 12px; 
+            line-height: 16px;   
+            margin-left: 3px;
+            margin-right: 3px;            
+        }
+        
+    }
+}
+
+@media (min-width: 768px) {
+    .board {
+        position: fixed;
+        width: 304px;
+        will-change: top, left;
+        top: 48px;
+        left: 270px;
+    }    
+}
+
+@media (min-width: 992px) {
+    .board {
+        position: fixed;
+        width: 304px;
+        will-change: top, left;
+        top: 48px;
+        left: 400px;
+    }    
+}
+
+@media (min-width: 1290px) {
+    .board {
+        position: fixed;
+        width: 304px;
+        will-change: top, left;
+        top: 48px;
+        left: 985px;
+    }    
+}
+
+
+
+
