+++ src/styles/pages/home.scss
@@ -49,7 +49,7 @@
             border-radius: 20px;
         }
 
-
+
