+++ src/styles/pages/mainboard.scss
@@ -74,7 +74,7 @@
 .workSpace{
   background-color: #0079BF;
   width: 100%;
-  height: 500px;
+  height: 90vh;
   padding-top: 30px;
   
   .textContent{
